+++ resolwe/storage/connectors/baseconnector.py
@@ -315,6 +315,14 @@ class BaseStorageConnector(metaclass=abc.ABCMeta):
         """
         raise NotImplementedError
 
+    @validate_url
+    def url(self, url: Union[str, PathLike]) -> Optional[str]:
+        """Return the URL to the given file in connector native notation.
+
+        :returns: the url to the given file in connector native notation.
+        """
+        raise NotImplementedError
+
     def __eq__(self, other):
         """Equality check.