+++ docs/CHANGELOG.rst
@@ -15,6 +15,8 @@ Added
 
 - Add handler ``resolve_url`` to listener
 - Add support for resolving urls in Python processes
+- Add ``url`` method to the ``BaseConnector`` class and override it in ``S3``
+  and ``local`` connectors
 
 
 ===================